Move 'too-recent' check after 'already-submitted*' checks
It'd be wrong to mark a change as being too recent, while the best
reason is that it was already submitted.

This means at least one more request to the build service (to see if it
was declined before) at first. But on the other hand, on later runs, we
won't check things for this change anymore as 'already-submitted*'
filters are permanent, while 'too-recent' has to be verified again and
again.