Improve check for request revision equality
A package revision can be in multiple formats:
a) the numeric revision
b) srcmd5 (hash of the unexpanded package contents)
c) xsrcmd5 (hash of the expanded package contents)

A request can contain any of those as source revision.

To check whether a particular package was already submitted, look at all of
those.